make other routes protected

+++ src/main/java/com/singhabraj/myFirstProject/service/JournalEntityService.java
@@ -43,11 +43,22 @@ public class JournalEntityService {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id, String userName){
-        User user= userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x->x.getId().equals(id));
-         userService.saveUser(user);
-         journalEntryRepository.deleteById(id);
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName){
+    boolean removed = false;
+       try{
+           User user= userService.findByUserName(userName);
+            removed = user.getJournalEntries().removeIf(x->x.getId().equals(id));
+           if(removed){
+               userService.saveUser(user);
+               journalEntryRepository.deleteById(id);
+           }
+
+       }catch(Exception e){
+           throw new RuntimeException("An error occurred while deleting the entry.",e);
+       }
+       return removed;
+
     }
 
 }
